JavaScript Exercise #2 - Student's Final Grade Average
Student's Final Grade Average

Write a program to determine the grades of the student.

Take the student's name, class and marks as input. Marks can be in the range of 1-100. Grading mechanism is as follow:
A - 80-100
B - 60-80
C - 40-60
D - <40

Determine what the grade is and print the output in a table format in HTML which will have student's name and grade.