JavaScript Tutorial Exercises
JavaScript Exercise #1 - Palindrome Words

Write a function to check if the given string is a palindrome or not. Create an input textbox to get the input string from the user and a button to check if it is. Show the results in alert and console. Also, there should be different js file and html file and link both using the script tag (Google this, you will get it easily).